+++ test/sql/misc.sql
@@ -663,13 +663,15 @@ begin;
   select * from emaj.emaj_verify_all();
 rollback;
 -- detection of tables altered as UNLOGGED
-begin;
-  alter table "phil's schema3"."myTbl2\" set unlogged;                        -- needs 9.5+
+begin;                                                                        -- needs 9.5+
+  alter table myschema1.mytbl4 set unlogged;                                  -- table from a rollbackable group
+  alter table "phil's schema3"."myTbl2\" set unlogged;                        -- table from an audit_only group
   select * from emaj.emaj_verify_all();
 rollback;
 -- detection of tables altered as WITH OIDS
 begin;
-  alter table "phil's schema3"."myTbl2\" set with oids;
+  alter table myschema1.mytbl4 set with oids;                                 -- table from a rollbackable group
+  alter table "phil's schema3"."myTbl2\" set with oids;                       -- table from an audit_only group
   select * from emaj.emaj_verify_all();
 rollback;
 -- detection of modified primary key
@@ -697,7 +699,7 @@ begin;
   alter table myschema1.mytbl1 add column newcol int;
   update emaj.emaj_relation set rel_kind = 'S' where rel_schema = 'myschema2' and rel_tblseq = 'mytbl1' and upper_inf(rel_time_range);
   alter table myschema1.mytbl4 drop constraint mytbl4_pkey;
-  alter table "phil's schema3"."myTbl2\" set with oids;
+  alter table myschema1.mytbl4 set with oids;
   select * from emaj.emaj_verify_all();
 rollback;